refactor: Move xmlToJson to Deprecated.js
