+++ lib/itoolkit.js
@@ -23,7 +23,6 @@ const { ProgramCall } = require('./ProgramCall');
 const { CommandCall } = require('./CommandCall');
 const { Connection } = require('./Connection');
 const { Toolkit } = require('./Toolkit');
-const { xmlToJson } = require('./utils');
 
 const {
   iPgm,
@@ -38,6 +37,7 @@ const {
   iCmd,
   iQsh,
   iSh,
+  xmlToJson,
 } = require('./Deprecated');
 
 module.exports = {
